+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Move all URL and file path / name / etc. functionality to `DocLocation` class
 
 ### 1.5.6